Minor refactoring of test package
This commit renames the package with unit tests to 'test',
like the parent folder. It also removes a typo in the name
of a global variable.
