+++ test/signatures_test.go
@@ -1,4 +1,4 @@
-package tests
+package test
 
 import (
 	"github.com/xlab-si/emmy/common"
